practice/chapter2/code/requirements.txt
# Chapter 2: 데이터 전처리와 특성 공학

# 기본 라이브러리
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
scikit-learn>=1.3

# 차원 축소
umap-learn>=0.5

# 텍스트 임베딩
sentence-transformers>=2.2

# Entity Embedding
torch>=2.0

# 환경 변수
python-dotenv>=1.0
